requirements.txt
streamlit
python-Levenshtein